Convert sampler to be runnable with pub global run
