+++ packages/snippets/bin/extract_sample.dart
