+++ utils/sampler/AUTHORS
@@ -1,0 +1,7 @@
+# Below is a list of people and organizations that have contributed
+# to this package. Names should be added to the list like so:
+#
+#   Name/Organization <email address>
+
+Google, LLC
+
